% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officerpack.R
\name{cf_example}
\alias{cf_example}
\title{cf_example}
\usage{
cf_example(...)
}
\description{
Simple way to create word document using loop of from t list.
Type lhtext and copy the template to R workspace and start writing.
}
\keyword{lhtext}
