% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officerpack.R
\name{fnote}
\alias{fnote}
\title{fnote}
\usage{
fnote(data)
}
\arguments{
\item{data}{list of define. ex<-c("A","C","B"); fnote(ex)= "ABC"}
}
\description{
convert and sorte list of define in footnote to one paragraph.
}
\examples{
fnote()

}
\keyword{fnote}
