% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officerpack.R
\name{l}
\alias{l}
\title{len}
\usage{
l(t)
}
\arguments{
\item{t}{ex: t[[1]] and the subsequent t is t[[len(t)]]. Convenience is you can insert t after the first t}
}
\description{
could be used in lhtext
}
\examples{
len()

}
\keyword{len}
