% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officerpack.R
\name{lhflex}
\alias{lhflex}
\title{lhflex}
\usage{
lhflex(table1, csv = "yes", bord = "yes", select = NULL,
  add.h = NULL, merge.all = "yes", size = 12, empty = NULL,
  cf = NULL, border = NULL, align = "center")
}
\arguments{
\item{csv}{If source = csv otherwise flextable tab}

\item{add.h}{Define additional header rows df<-data.frame(row1=c("",rep("median (CV%)",4),row2=c("Inches","Inches","Inches"#',"Inches","Species") unit=c("mg/mL","inch"," ",       " "," "))) then add.h=df}

\item{cf}{Conditional formatting see cf_example}

\item{border}{Border list("vi:dashed:black:header","vo:dashed:black:body","ho:dashed:black:body",etc.)}

\item{lst}{List of header example lst=c(Mean="mean",animal="dog")}
}
\description{
Simple way to create word document using loop of from t list.
Type lhtext and copy the template to R workspace and start writing.
}
\keyword{lhflex}
