% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officerpack.R
\name{lhflextab}
\alias{lhflextab}
\title{lflextab}
\usage{
lhflextab(table1, csv = "yes", bord = "yes", select = NULL,
  add.h = NULL, merge.all = "yes", size = 12, empty = NULL,
  cf = NULL, border = NULL, align = "center")
}
\arguments{
\item{csv}{If source = csv otherwise flextable tab}

\item{add.h}{Define additional header rows Ex: df<-data.frame(row1=c("",rep("median (CV%)",4),row2=c("Inches","Inches","Inches"#',"Inches","Species") unit=c("mg/mL","inch"," ",       " "," "))) then add.h=df}

\item{cf}{conditional formatting. should be list("i=x, j=y :(format abbreviation) format: col=color (:colred),ita=italic (:ita),bol=bold (:bol), bg=background (:bgred).  Conditional statement = i=~colname >or< or == values, j=~col1+col2". mv=vertical (merge identical value), mh=horizontal (merge identical value); ma=at (merge cells regardless values). ex. mv=list("j=1 or j=~colname :mv"); ma = c.}

\item{border}{Border list("vi:dashed:black:header","vo:dashed:black:body","ho:dashed:black:body",etc.)}

\item{lst}{List of header example lst=c(Mean="mean",animal="dog")}
}
\description{
Create doc for word document using Officer.
}
\examples{

       
}
\keyword{officer.report.template}
