% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officerpack.R
\name{lhfootnote}
\alias{lhfootnote}
\title{lhfootnote}
\usage{
lhfootnote(list)
}
\arguments{
\item{list}{List of define ex. list=c(def1,def2,etc). output should be insert in txt and set font size}
}
\description{
convert and sorte list of define in footnote to one paragraph.
}
\examples{
lhfootnote()

}
\keyword{fnote}
