% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officerpack.R
\name{lhstartdoc}
\alias{lhstartdoc}
\title{officer.report.template}
\usage{
lhstartdoc(temp = "c:/lhtemplate/stylereport.docx", TFL = T)
}
\arguments{
\item{TFL}{FAULT for no table and figure lists}
}
\description{
Create doc for word document using Officer.
}
\keyword{officer.report.template}
