% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officerpack.R
\name{lhtext_example}
\alias{lhtext_example}
\title{lhtext_example}
\usage{
lhtext_example(...)
}
\description{
Simple way to create word document using loop of from t list.
Type lhtext and copy the template to R workspace and start writing.
}
\keyword{lhtext}
